requirements.txt
Django>=4.1
djangorestframework>=3.14.0
smbprotocol>=1.4
flake8
black